move register forward global to opc ua server stack
