+++ src/OpcUaStackServer/ServiceSetApplication/RegisterForwardGlobalRequest.h
@@ -1,5 +1,5 @@
 /*
-   Copyright 2017-2019 Kai Huebl (kai@huebl-sgh.de)
+   Copyright 2017-2020 Kai Huebl (kai@huebl-sgh.de)
 
    Lizenziert gemäß Apache Licence Version 2.0 (die „Lizenz“); Nutzung dieser
    Datei nur in Übereinstimmung mit der Lizenz erlaubt.
@@ -15,33 +15,33 @@
    Autor: Kai Huebl (kai@huebl-sgh.de)
  */
 
-#ifndef __OpcUaStackCore_RegisterForwardGlobalResponse_h__
-#define __OpcUaStackCore_RegisterForwardGlobalResponse_h__
+#ifndef __OpcUaStackServer_RegisterForwardGlobalRequest_h__
+#define __OpcUaStackServer_RegisterForwardGlobalRequest_h__
 
-#include "OpcUaStackCore/Base/Object.h"
-#include "OpcUaStackCore/BuildInTypes/OpcUaStatusCode.h"
+#include "OpcUaStackCore/ServiceSetApplication/ForwardGlobalSync.h"
 
-namespace OpcUaStackCore
+namespace OpcUaStackServer
 {
 
-	class DLLEXPORT RegisterForwardGlobalResponse
-	: public Object
+	class DLLEXPORT RegisterForwardGlobalRequest
+	: public OpcUaStackCore::Object
 	{
 	  public:
-		typedef boost::shared_ptr<RegisterForwardGlobalResponse> SPtr;
+		typedef boost::shared_ptr<RegisterForwardGlobalRequest> SPtr;
 
-	    RegisterForwardGlobalResponse(void);
-		virtual ~RegisterForwardGlobalResponse(void);
+		RegisterForwardGlobalRequest(void);
+		virtual ~RegisterForwardGlobalRequest(void);
 
-		void statusCode(const OpcUaStatusCode statusCode);
-		OpcUaStatusCode statusCode(void) const;
+		void forwardGlobalSync(OpcUaStackCore::ForwardGlobalSync::SPtr forwardGlobalSync);
+		OpcUaStackCore::ForwardGlobalSync::SPtr forwardGlobalSync(void);
 
 		bool opcUaBinaryEncode(std::ostream& os) const;
 		bool opcUaBinaryDecode(std::istream& is);
 
 	  private:
-		OpcUaStatusCode statusCode_;
+		OpcUaStackCore::ForwardGlobalSync::SPtr forwardGlobalSync_;
 	};
+
 }
 
 #endif
